[DOP-9718] - move module etl_entitites.hwm -> etl_entitites.old_hwm (#27)
