+++ docs/changelog/next_release/27.breaking.rst
@@ -1,0 +1,1 @@
+Move module from ``etl_entitites.hwm`` to ``etl_entities.old_hwm`` and mark these classes as deprecated
