+++ docs/old_hwm/date_hwm.rst
@@ -1,8 +1,8 @@
-Datetime HWM
+Date HWM
 =================================================================
-.. currentmodule:: etl_entities.hwm.datetime_hwm
+.. currentmodule:: etl_entities.old_hwm.date_hwm
 
 
-.. autoclass:: DateTimeHWM
+.. autoclass:: DateHWM
     :members: name, qualified_name, set_value, dict, json, copy, serialize, deserialize, covers, update
     :special-members: __bool__, __add__, __sub__, __eq__, __lt__
