+++ docs/old_hwm/datetime_hwm.rst
@@ -1,8 +1,8 @@
-Date HWM
+Datetime HWM
 =================================================================
-.. currentmodule:: etl_entities.hwm.date_hwm
+.. currentmodule:: etl_entities.old_hwm.datetime_hwm
 
 
-.. autoclass:: DateHWM
+.. autoclass:: DateTimeHWM
     :members: name, qualified_name, set_value, dict, json, copy, serialize, deserialize, covers, update
     :special-members: __bool__, __add__, __sub__, __eq__, __lt__
