+++ docs/old_hwm/int_hwm.rst
@@ -1,6 +1,6 @@
 Integer HWM
 =================================================================
-.. currentmodule:: etl_entities.hwm.int_hwm
+.. currentmodule:: etl_entities.old_hwm.int_hwm
 
 
 .. autoclass:: IntHWM
