+++ etl_entities/old_hwm/file_list_hwm.py
@@ -20,9 +20,9 @@ from typing import FrozenSet, Iterable, List
 
 from pydantic import Field, validator
 
-from etl_entities.hwm.file_hwm import FileHWM
-from etl_entities.hwm.hwm_type_registry import register_hwm_type
 from etl_entities.instance import AbsolutePath, RelativePath
+from etl_entities.old_hwm.file_hwm import FileHWM
+from etl_entities.old_hwm.hwm_type_registry import register_hwm_type
 
 FileListType = FrozenSet[RelativePath]
 
@@ -30,6 +30,8 @@ FileListType = FrozenSet[RelativePath]
 @register_hwm_type("file_list")
 class FileListHWM(FileHWM[FileListType, List[str]]):
     """File List HWM type
+
+    .. deprecated:: 2.0.0
 
     Parameters
     ----------
@@ -58,7 +60,7 @@ class FileListHWM(FileHWM[FileListType, List[str]]):
 
         folder = RemoteFolder(name="/absolute/path", instance="ftp://ftp.server:21")
 
-        hwm = FileListHWM(
+        old_hwm = FileListHWM(
             source=folder,
             value=["some/path", "another.file"],
         )
@@ -114,11 +116,11 @@ class FileListHWM(FileHWM[FileListType, List[str]]):
 
             from etl_entities import FileListHWM
 
-            hwm = FileListHWM(value=["some/file.py", "another.file"], ...)
-            assert hwm.serialize_value() == ["some/file.py", "another.file"]
+            old_hwm = FileListHWM(value=["some/file.py", "another.file"], ...)
+            assert old_hwm.serialize_value() == ["some/file.py", "another.file"]
 
-            hwm = FileListHWM(value=[], ...)
-            assert hwm.serialize_value() == []
+            old_hwm = FileListHWM(value=[], ...)
+            assert old_hwm.serialize_value() == []
         """
 
         return sorted(os.fspath(item) for item in self.value)
@@ -178,10 +180,10 @@ class FileListHWM(FileHWM[FileListType, List[str]]):
 
             from etl_entities import FileListHWM
 
-            hwm = FileListHWM(value=["some/path.py"], ...)
+            old_hwm = FileListHWM(value=["some/path.py"], ...)
 
-            assert hwm.covers("some/path.py")  # path in HWM
-            assert not hwm.covers("another/path.py")  # path not in HWM
+            assert old_hwm.covers("some/path.py")  # path in HWM
+            assert not old_hwm.covers("another/path.py")  # path not in HWM
         """
 
         return value in self
@@ -206,18 +208,18 @@ class FileListHWM(FileHWM[FileListType, List[str]]):
 
             from etl_entities import FileListHWM
 
-            hwm = FileListHWM(value=["some/existing_path.py"], ...)
+            old_hwm = FileListHWM(value=["some/existing_path.py"], ...)
 
             # new paths are appended
-            hwm.update("some/new_path.py")
-            assert hwm.value == [
+            old_hwm.update("some/new_path.py")
+            assert old_hwm.value == [
                 "some/existing_path.py",
                 "some/new_path.py",
             ]
 
             # existing paths do nothing
-            hwm.update("some/existing_path.py")
-            assert hwm.value == [
+            old_hwm.update("some/existing_path.py")
+            assert old_hwm.value == [
                 "some/existing_path.py",
                 "some/new_path.py",
             ]
@@ -245,11 +247,11 @@ class FileListHWM(FileHWM[FileListType, List[str]]):
 
             from etl_entities import FileListHWM
 
-            hwm = FileListHWM(value=["some/path.py"], ...)
-            assert hwm  # same as bool(hwm.value)
+            old_hwm = FileListHWM(value=["some/path.py"], ...)
+            assert old_hwm  # same as bool(old_hwm.value)
 
-            hwm = FileListHWM(value=[], ...)
-            assert not hwm
+            old_hwm = FileListHWM(value=[], ...)
+            assert not old_hwm
         """
 
         return bool(self.value)
@@ -390,9 +392,11 @@ class FileListHWM(FileHWM[FileListType, List[str]]):
 
             from etl_entities import FileListHWM, Folder, AbsolutePath
 
-            hwm = FileListHWM(value=["some/path"], source=Folder(name="/absolute/path", ...), ...)
+            old_hwm = FileListHWM(
+                value=["some/path"], source=Folder(name="/absolute/path", ...), ...
+            )
 
-            assert abs(hwm) == frozenset(AbsolutePath("/absolute/path/some/path"))
+            assert abs(old_hwm) == frozenset(AbsolutePath("/absolute/path/some/path"))
         """
 
         return frozenset(self.source / item for item in self.value)
@@ -413,10 +417,12 @@ class FileListHWM(FileHWM[FileListType, List[str]]):
 
             from etl_entities import FileListHWM, Folder, AbsolutePath
 
-            hwm = FileListHWM(value=["some/path"], source=Folder(name="/absolute/path", ...), ...)
+            old_hwm = FileListHWM(
+                value=["some/path"], source=Folder(name="/absolute/path", ...), ...
+            )
 
-            assert "some/path" in hwm
-            assert "/absolute/path/some/path" in hwm
+            assert "some/path" in old_hwm
+            assert "/absolute/path/some/path" in old_hwm
         """
 
         if not isinstance(item, os.PathLike):
@@ -429,7 +435,7 @@ class FileListHWM(FileHWM[FileListType, List[str]]):
 
         Params
         -------
-        other : :obj:`hwmlib.hwm.file_list_hwm.FileListHWM`
+        other : :obj:`hwmlib.old_hwm.file_list_hwm.FileListHWM`
 
         Returns
         --------
